+++ services/director-v2/src/simcore_service_director_v2/core/dynamic_services_settings/__init__.py
@@ -1,8 +1,8 @@
 from pydantic import Field
 from settings_library.base import BaseCustomSettings
 
-from .dynamic_sidecar_settings import DynamicSidecarSettings
 from .scheduler import DynamicServicesSchedulerSettings
+from .sidecar import DynamicSidecarSettings
 
 
 class DynamicServicesSettings(BaseCustomSettings):
