+++ services/director-v2/src/simcore_service_director_v2/modules/node_rights.py
@@ -10,9 +10,7 @@ from redis.asyncio.lock import Lock
 from servicelib.redis import RedisClientSDK
 from settings_library.redis import RedisDatabase, RedisSettings
 
-from ..core.dynamic_services_settings.dynamic_sidecar_settings import (
-    DynamicSidecarSettings,
-)
+from ..core.dynamic_services_settings.sidecar import DynamicSidecarSettings
 from ..core.errors import ConfigurationError, NodeRightsAcquireError
 
 DockerNodeId = str
