+++ services/director-v2/tests/unit/test_core_settings.py
@@ -9,7 +9,7 @@ from models_library.basic_types import LogLevel
 from pydantic import ValidationError
 from pytest_simcore.helpers.typing_env import EnvVarsDict
 from settings_library.r_clone import S3Provider
-from simcore_service_director_v2.core.dynamic_services_settings.dynamic_sidecar_settings import (
+from simcore_service_director_v2.core.dynamic_services_settings.sidecar import (
     DynamicSidecarSettings,
     EnvoyLogLevel,
     RCloneSettings,
